Allow passing buttons to popup and show them
Also fire event back for resolving promise in popup with name of button clicked

+++ src/libs/recorder/recordingsManager.ts
@@ -142,20 +142,22 @@ export default class RecordingsManager {
     const manifestStream = fs.createWriteStream(tmpOutFolder + "/manifest.txt", { flags: "a" });
     const popupName = "clipVideo";
 
-    Notifications.popup(popupName, "Clipping Your Video", undefined, () => {
-      Notifications.popup(popupName, "Cancelling Processing Of Your Video");
+    Notifications.popup(popupName, "Clipping Your Video", { loader: true, showCancel: true }).then((action) => {
+      if (action == "cancel") {
+        Notifications.popup(popupName, "Cancelling Processing Of Your Video");
 
-      // Stop ffmpeg and destroy manifestStream
-      ffmpeg.kill();
-      manifestStream.destroy();
+        // Stop ffmpeg and destroy manifestStream
+        ffmpeg.kill();
+        manifestStream.destroy();
 
-      // Remove associated files/folders if they exist
-      PathHelper.removeDir(tmpOutFolder);
-      PathHelper.removeFile(clipOutPath);
+        // Remove associated files/folders if they exist
+        PathHelper.removeDir(tmpOutFolder);
+        PathHelper.removeFile(clipOutPath);
 
-      // When FFmpeg is closed, popup is also deleted below, but FFmpeg won't always
-      // be open when user is cancelling so also delete it here just incase.
-      Notifications.deletePopup(popupName);
+        // When FFmpeg is closed, popup is also deleted below, but FFmpeg won't always
+        // be open when user is cancelling so also delete it here just incase.
+        Notifications.deletePopup(popupName);
+      }
     });
 
     // Create clips from video.
